+++ shell/Makefile
@@ -6,3 +6,9 @@ clean :
 
 test :
 	../test/run.sh host_shell
+
+CC = gcc
+CFLAGS = -std=c99 -Wall -Werror
+
+shell : ../src/sys_calls.o shell.o
+	$(CC) $(LDFLAGS) -o $@ $^
