--- src/apps.h
@@ -1,7 +1,0 @@
-#ifndef EDUOS_APPS_H
-#define EDUOS_APPS_H
-
-extern void shell();
-
-#endif /* EDUOS_APPS_H */
-
